# Conflicts:
#	mainapp/src/main/res/layout/o_activity_two_textviews.xml

+++ mainapp/src/main/java/biyaniparker/com/parker/view/homeadmin/orderdispatch/RecentDispatchDetailView.java
@@ -202,7 +202,7 @@ public class RecentDispatchDetailView extends AppCompatActivity {
                 }
                 //linear.addView(linear);
             }
-            txtTPrice.setText("Total : " + (totalQnty * Integer.parseInt(detailBean.getConsumerPrice())));
+            txtTPrice.setText("Total : " + (totalQnty * Float.parseFloat(detailBean.getConsumerPrice())));
             mainlinearDetails.addView(convertView);
